if delete category, change post with that category to uncategorize option. and also can't delete uncategorize

+++ app/Http/Controllers/CategoryController.php
@@ -106,6 +106,9 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        if ($id == 1)
+            return abort(403);
+
         $category = Category::findOrFail($id);
         $category->delete();
